introducing new enum ColorPickerWidgetType to differentiate between hover color picker and standalone color picker
